Disabled refreshing an output panel on every window activation - it's annoying.
I don't want to include state management into the code yet, which is required to implement this feature in a way that Sublime Text basic console works. So this feature would work in a way to print out the very complete history to a very last window where the plugin did printed out it's output. In all other windows would be outdated history provided, that at its turn would be updated on a plugin call with that window being active at that time.

This is incomplete solution, but it should fix a significant numbers of use cases while not being annoying.