Ignorance of global parameters correction
Correct a bug where the openai_panel command would give the default chatGPT url for a model in the case where the url is not specified for each model individually.
With this modification, it will take first the 'url' setting that is set globally in the setting file.
Note: did the same think for the 'token' setting.